Adding the 1c subduction benchmark as a test with direct solvers for the moment.
This uses fieldsplits to separate out the portion of the domain with prognostic fields from those without, however petsc
automatically creates conjugate fieldsplits for the other sections and does something unclear with them leading to some strange
results.  Hence, I've added the ability to take control of these conjugate fieldsplits by adding parental and sibling inheritance
to the creation of index sets.  Now indices that are described as belonging to a fieldsplit that don't belong to its parent are
removed (with a warning).  Similarly indices that already belong to a sibling fieldsplit are removed - this is necessary because
with region id based fieldsplits you always end up with overlapping index sets for continuous elements.  These changes now mean it
is possible to request a fieldsplit (or a null space, which is also described using index sets) without specifying any fields and it
will default to the parental index set minus any siblings that have already been assigned.  As ever, siblings are taken in the order
they appear in the options file.

Other new features...
 - bilinear forms now have the option to ident any zero rows found - for instance in the non prognostic regions these can just have
preonly-jacobi thrown at them.
 - correcting the comments for meshes from file
 - correcting a bug when specifying index sets from system function spaces that are not mixed
 - adding the option of viewing an IS
 - can return the fields_.size() from a system - necessary to detemine if its mixed or not