Split up the initialization of fields and coefficient expressions from that of coefficient functions.
The latter requires an eval and if it is dependent on fields or expressions in other systems it is likely to return incorrect
values.

Field initial condition eval is also split out into a final step so that the order of initialization (that matters to
interdependencies from the users perspective) is now:

loop over systems:
  loop over coefficient functions
loop over systems:
  loop over field initial conditions

In other words, coefficient functions should not be made to depend on field initial conditions and cannot depend on the values of
coefficient functions defined lower down in the tfml file.