Another iteration working on extracting IS's.
This iteration undos a lot of the previous work but extracts general components out of the class sytem.  This could be improved in
the mvc_parallel branch once merged.

This fixes the long standing bug with cap values.

Next step null spaces but these changes will also be very important in optimizing output.