Parallel null spaces (interacting with fieldsplits)  now seem to be working.
This involved a lot of consolidation of the IS code for fieldsplits with those for values (null vectors and bounds).

Additionally to test this some functionality was added call zero_points that manipulate the solution vector (this again uses the
capability of pulling component ISs out for any subfunctionspace).

The resulting test solves the same poisson problem with neumann bcs repeatedly at different levels of the fieldsplit.  As this (and
the other mms poisson test) is much better than the previous test, I'm deleting it.