A reimplementation of ReferencePoints (now called ReferencePoint).
These no longer use big springs but are instead implemented as overloaded pointwise DirichletBC so that they can be assembled within
the SystemAssembler.  This has obvious advantages but makes the process of finding the dof more challenging as it relies on floating
point comparison (unlike the old way) within a SubDomain.  I've tried to mitigate the problems of using this on subcomponents by
forcing each component to have a separate ReferencePoint.  This complication is hidden from the user and the interface has not changed.