A few changes required by dolfin 1.5.0.
functionspace.cell_dofs now returns local not global indices so return this to the old behaviour by converting back to global
afterwards.

vector.operator[] now takes in local indices rather than global so replace calls to this with get, which remains global.

The great news is that dg parallel assembly now works so set up a test of this by modifying some mms poisson tests.  The dg version
of this (which has facet integrals) requires a new option specifying that the ghost mode should use shared facets so have added
optiosn to the schema to allow this (and while I'm at it allow_extrapolation) to be specified.  The drt version (which does not have
facet integrals) does not require this option.