When calculating index sets for fieldsplits communicate off-process nodes to their owners to make sure they include them too.
This solution is overkill as it happens before the indices are restricted by parents or siblings but doing anything more
sophisticated will require carrying around ownership information for the parent indices.

Adding a test of this that runs the stokes mms case but on a larger domain that is then fieldsplit by region and run in parallel on various
'problem' partitionings.  To do this a little debugging functionality has been added to allow users to specify a map from process
rank to region_ids.  This is not recommended for production runs as it will read the entire mesh in serial before partitioning.