Introducing the ability to specify the solution order of the problem.
By default this still does what the old system did - solve systems by system and then by nonlinear_solver.  However if the the
solution_order option is activated for any of the three solve locations then this is overridden and instead the user can specify
what to solve, when and in what order.

This incorporates and replaces the nonlinear_systems option by allowing solvers to be called within an iterative loop.  Successive
iterative loops can be nested or called sequentially.

This has all been done by introducing a new class (and some associated classes): the SystemsSolverBucket.  This contains lists of
(generic) solvers, which are either references to standard systems nonlinear_solvers or references to nested SystemsSolverBuckets and are
solved sequentially.

By default the SystemsSolverBuckets are populated using the solve locations underneath the system nonlinear_solvers, resulting in
the previous behaviour.  However if the user activates the solution_order option then they are populated by what the user specifies.

All solves now take place through the SystemsSolverBuckets, this effectively short-circuits the hierarchical structure as the Bucket
now manages the SystemsSolverBuckets, which in turn manage the SolverBuckets, bypassing the SystemBuckets, which were previously the
intermediary level.

Use updatetfml and updateshml to modify options files to the new schema.  All tests have been updated.  Additionally one bug in
updateshml was fixed for renaming of functionals to prevent erroneous modification of previously fixed files.