Adding some rudimentary running form checkpoints capabilities to tfsimulationharness.
A new option --run-from-checkpoint [index] has been added, that will look for suitable checkpoint files and restart simulations
(without rebuilding) from them.  Small changes to options are allowed but must not require a rebuild.

A helper function _self.numbercheckpoints is provided to help find files for variable computation if restarted from checkpoint.

It is also now possible to provide required files as python dictionaries where key:value pairs rename the file as it is copied - so
for instance, dependency simulations can take in the same input even if it has come from a checkpoint.

This functionality is now tested during shorttests through the projection_timedependent test.