This seems to have got xdmf checkpointing working.
Functions have to be checkpointed individually at the FunctionBucket level rather than the SystemBucket level.  This has the
advantage of allowing some fields to started from checkpoint while others in the same system are not.  However it is a bit of a
change from how we did it before and FunctionBuckets now have FunctionAssigners (as well as the already existing component IS's...
overkill!) to allow us to do it.

As a result of these changes InitialConditionExpression is dead and initialize_over_regions now defaults to 0 when no other option
is present.

Also fixing an error in the simulation harness schema so that it now enforces python3.